# docs/api/openapi.yaml
openapi: 3.2.1
info:
  title: School Report API
  version: 1.0.0
  description: |
    REST API for the School Report MVP for Indonesian elementary schools.
    One deployed application belongs to exactly one school. The authenticated
    session determines school context; clients MUST NOT use schoolId for authorization.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /report-cards:
    get:
      tags: [Reports]
      operationId: listReportCards
      parameters:
        - { in: query, name: academicYearId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: semesterId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: classId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: studentId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/ReportCardStatus' } }
      responses:
        '200': { description: Reports, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedReportCards' } } } }
  /report-cards/generate:
    post:
      tags: [Reports]
      operationId: generateReportCard
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/GenerateReportCardRequest' } } }
      responses:
        '201': { description: Generated snapshot, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCardDetail' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /report-cards/{reportCardId}:
    get:
      tags: [Reports]
      operationId: getReportCard
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      responses:
        '200': { description: Report card, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCardDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /report-cards/{reportCardId}/review:
    post:
      tags: [Reports]
      operationId: reviewReportCard
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      responses:
        '200': { description: Reviewed, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCardDetail' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /report-cards/{reportCardId}/lock:
    post:
      tags: [Reports]
      operationId: lockReportCard
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      responses:
        '200': { description: Locked, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCardDetail' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /report-cards/{reportCardId}/publish:
    post:
      tags: [Reports]
      operationId: publishReportCard
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      responses:
        '200': { description: Published, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCardDetail' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /report-cards/{reportCardId}/revision:
    post:
      tags: [Reports]
      operationId: createReportRevision
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateRevisionRequest' } } }
      responses:
        '201': { description: New revision, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCardDetail' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /report-cards/{reportCardId}/pdf:
    get:
      tags: [Reports]
      operationId: downloadReportPdf
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      responses:
        '200':
          description: PDF
          content: { application/pdf: { schema: { type: string, format: binary } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /report-cards/{reportCardId}/subjects/{subjectId}/description:
    patch:
      tags: [Reports]
      operationId: updateSubjectDescription
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }, { $ref: '#/components/parameters/SubjectId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateSubjectDescriptionRequest' } } }
      responses:
        '200': { description: Updated description, content: { application/json: { schema: { $ref: '#/components/schemas/SubjectDescriptionResult' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /report-cards/{reportCardId}/subjects/{subjectId}/description/reset:
    post:
      tags: [Reports]
      operationId: resetSubjectDescription
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }, { $ref: '#/components/parameters/SubjectId' }]
      responses:
        '200': { description: Reset to generated description, content: { application/json: { schema: { $ref: '#/components/schemas/SubjectDescriptionResult' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /report-cards/{reportCardId}/completeness:
    patch:
      tags: [Reports]
      operationId: updateReportCompleteness
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateCompletenessRequest' } } }
      responses:
        '200': { description: Updated completeness, content: { application/json: { schema: { $ref: '#/components/schemas/CompletenessResult' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /report-cards/{reportCardId}/extracurriculars:
    get:
      tags: [Reports]
      operationId: listExtracurriculars
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      responses:
        '200': { description: Extracurricular entries, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ExtracurricularEntry' } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Reports]
      operationId: createExtracurricular
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExtracurricularRequest' } } }
      responses:
        '201': { description: Created entry, content: { application/json: { schema: { $ref: '#/components/schemas/ExtracurricularEntry' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /report-cards/{reportCardId}/extracurriculars/{entryId}:
    patch:
      tags: [Reports]
      operationId: updateExtracurricular
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }, { $ref: '#/components/parameters/ExtracurricularEntryId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExtracurricularRequest' } } }
      responses:
        '200': { description: Updated entry, content: { application/json: { schema: { $ref: '#/components/schemas/ExtracurricularEntry' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Reports]
      operationId: deleteExtracurricular
      parameters: [{ $ref: '#/components/parameters/ReportCardId' }, { $ref: '#/components/parameters/ExtracurricularEntryId' }]
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: school_report_session
  parameters:
...
    SubjectId: { in: path, name: subjectId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
...
    ReportCardId: { in: path, name: reportCardId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    ExtracurricularEntryId: { in: path, name: entryId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
...
  responses:
...
    Forbidden: { description: Not authorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Conflict: { description: Domain conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    ValidationError: { description: Validation/domain error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  schemas:
    UUID: { type: string, format: uuid }
...
    DateTime: { type: string, format: date-time }
...
    ReportCardStatus: { type: string, enum: [DRAFT, REVIEW, LOCKED, PUBLISHED, REVISION] }
    AchievementStatus: { type: string, enum: [ACHIEVED, NOT_ACHIEVED, NOT_ASSESSED] }
...
    ReportCardSubject:
      type: object
      required: [id, reportCardId, subjectId, subjectName, finalScore, achievement]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        reportCardId: { $ref: '#/components/schemas/UUID' }
        subjectId: { $ref: '#/components/schemas/UUID' }
        subjectName: { type: string }
        finalScore: { type: number, minimum: 0, maximum: 100 }
        kktpThreshold: { type: number, minimum: 0, maximum: 100, nullable: true }
        achievement: { $ref: '#/components/schemas/AchievementStatus' }
        description: { type: string, nullable: true }
        descriptionSource: { type: string, enum: [AUTO, MANUAL] }
        snapshotJson: { type: object, additionalProperties: true, nullable: true }
    ReportCard:
      type: object
      required: [id, studentId, academicYearId, semesterId, classId, version, status]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        studentId: { $ref: '#/components/schemas/UUID' }
        academicYearId: { $ref: '#/components/schemas/UUID' }
        semesterId: { $ref: '#/components/schemas/UUID' }
        classId: { $ref: '#/components/schemas/UUID' }
        version: { type: integer, minimum: 1 }
        status: { $ref: '#/components/schemas/ReportCardStatus' }
        generatedAt: { allOf: [{ $ref: '#/components/schemas/DateTime' }], nullable: true }
        reviewedAt: { allOf: [{ $ref: '#/components/schemas/DateTime' }], nullable: true }
        lockedAt: { allOf: [{ $ref: '#/components/schemas/DateTime' }], nullable: true }
        publishedAt: { allOf: [{ $ref: '#/components/schemas/DateTime' }], nullable: true }
    ReportCardDetail:
      allOf:
        - $ref: '#/components/schemas/ReportCard'
        - type: object
          required: [subjects]
          properties:
            subjects: { type: array, items: { $ref: '#/components/schemas/ReportCardSubject' } }
            cocurricularDescription: { type: string, nullable: true }
            homeroomNotes: { type: string, nullable: true }
            sickDays: { type: integer, minimum: 0 }
            permissionDays: { type: integer, minimum: 0 }
            unexcusedDays: { type: integer, minimum: 0 }
            extracurriculars: { type: array, items: { $ref: '#/components/schemas/ExtracurricularEntry' } }
    GenerateReportCardRequest:
      type: object
      required: [studentId, academicYearId, semesterId]
      properties:
        studentId: { $ref: '#/components/schemas/UUID' }
        academicYearId: { $ref: '#/components/schemas/UUID' }
        semesterId: { $ref: '#/components/schemas/UUID' }
    CreateRevisionRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 1 }
    UpdateSubjectDescriptionRequest:
      type: object
      required: [description]
      properties:
        description: { type: string, minLength: 1, maxLength: 2000 }
    SubjectDescriptionResult:
      type: object
      required: [id, subjectId, descriptionSource]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        subjectId: { $ref: '#/components/schemas/UUID' }
        description: { type: string, nullable: true }
        descriptionSource: { type: string, enum: [AUTO, MANUAL] }
    UpdateCompletenessRequest:
      type: object
      properties:
        cocurricularDescription: { type: string, nullable: true }
        homeroomNotes: { type: string, nullable: true }
        sickDays: { type: integer, minimum: 0 }
        permissionDays: { type: integer, minimum: 0 }
        unexcusedDays: { type: integer, minimum: 0 }
    CompletenessResult:
      type: object
      required: [reportCardId, sickDays, permissionDays, unexcusedDays]
      properties:
        reportCardId: { $ref: '#/components/schemas/UUID' }
        cocurricularDescription: { type: string, nullable: true }
        homeroomNotes: { type: string, nullable: true }
        sickDays: { type: integer, minimum: 0 }
        permissionDays: { type: integer, minimum: 0 }
        unexcusedDays: { type: integer, minimum: 0 }
    ExtracurricularEntry:
      type: object
      required: [id, reportCardId, name, predicate, sortOrder]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        reportCardId: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        predicate: { type: string, enum: [A, B, C, D] }
        description: { type: string, nullable: true }
        sortOrder: { type: integer, minimum: 0 }
    ExtracurricularRequest:
      type: object
      required: [name, predicate]
      properties:
        name: { type: string, minLength: 1 }
        predicate: { type: string, enum: [A, B, C, D] }
        description: { type: string, nullable: true }
...
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    PaginatedReportCards: { type: object, required: [data, meta], properties: { data: { type: array, items: { $ref: '#/components/schemas/ReportCard' } }, meta: { $ref: '#/components/schemas/PaginationMeta' } } }
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - AUTH_INVALID_CREDENTIALS
            - AUTH_SESSION_EXPIRED
            - AUTH_FORBIDDEN
            - VALIDATION_ERROR
            - RESOURCE_NOT_FOUND
            - RESOURCE_CONFLICT
            - STUDENT_ALREADY_ENROLLED
            - INVALID_ACADEMIC_YEAR
            - INVALID_SEMESTER
            - INVALID_TEACHER_ASSIGNMENT
            - INVALID_WEIGHT_TOTAL
            - GRADING_SCHEME_PUBLISHED
            - ASSESSMENT_SCORE_EXCEEDED_MAX
            - ASSESSMENT_INCOMPLETE
            - REPORT_NOT_READY
            - REPORT_ALREADY_PUBLISHED
            - REPORT_NOT_LOCKED
            - REPORT_REVISION_REQUIRED
            - IMPORT_INVALID_FORMAT
            - IMPORT_VALIDATION_FAILED
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }